+++ sklearn_extra/robust/__init__.py
@@ -3,9 +3,12 @@ from sklearn_extra.robust.robust_weighted_estimator import (
     RobustWeightedKMeans,
     RobustWeightedRegressor,
 )
+from sklearn_extra.robust.mean_estimators import huber, make_huber_metric
 
 __all__ = [
     "RobustWeightedClassifier",
     "RobustWeightedKMeans",
     "RobustWeightedRegressor",
+    "huber",
+    "make_huber_metric",
 ]