+++ doc/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ----------
 
+- Add `make_huber_metric` which transform a non-robust to a robust metric using
+    Huber estimator.
 - Add a stopping criterion and parameter tuning heuristic for Huber robust mean
     estimator.
 - Add `CLARA` (Clustering for Large Applications) which extends k-medoids to